Moved to stack-based VM to fix drop/store/tee_local

+++ libwasmint/interpreter/ByteCode.h
@@ -75,7 +75,7 @@ namespace wasmint {
         }
 
         void appendOpcode(ByteOpcode opcode) {
-            append((uint16_t) opcode);
+            append((uint32_t) opcode);
         }
 
         template<typename T>
